pyproject.toml
[project]
name = "foxhound"
version = "0.1.0"
description = "Local, privacy-first document search and analysis pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "httpx>=0.28.1",
    "pymupdf>=1.27.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.2",
]
